requirements.txt
Flask==3.0.0
Pillow==10.0.0
numpy==1.26.0
matplotlib==3.8.0
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.0
pytest-cov==4.1.0